requirements.txt
asgiref==3.7.2
black==23.3.0
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.8
Django==4.2.2
django-taggit==4.0.0
djlint==1.30.2
EditorConfig==0.12.3
flake8==6.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.14.8
json5==0.9.14
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.1
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
